레시피, 저장, 플랜 도메인 리팩토링

+++ src/main/java/imwhs/eatz_server/dto/rating/RatingSummaryDto.java
@@ -16,12 +16,12 @@ public class RatingSummaryDto {
     /**
      *  레시피에 달린 평가 수
      */
-    private int ratingCount;
+    private Long ratingCount;
 
     /**
      * 레시피에 달린 평가들의 평균 점수
      * 1~5점 사이의 자연수 값이며, 사용자들이 남긴 평가 점수의 평균입니다.
      */
-    private double averageRatingScore;
+    private Double averageRatingScore;
 
 }
